refactor: tirando a responsabilidade de printar do controle e colocando em UI
